Update consul_sd example
