+++ examples/consul_sd/notif/pyproject.toml
@@ -8,7 +8,7 @@ authors = ["Guillaume Gauvrit <guillaume@gauvr.it>"]
 python = "^3.9"
 starlette = "^0.17.0"
 uvicorn = "^0.15.0"
-blacksmith = "^0.5.0"
+blacksmith = "^0.12.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
